+++ changelog.md
@@ -1,3 +1,7 @@
+# 09-03-2025 :
+- added `get_elements_by_class, get_elements_by_tag, get_elements_by_query, get_elements_by_name, _to_element, make_element` functions
+- added `get_elements_by_class, get_elements_by_tag, get_elements_by_query, get_elements_by_name, _to_element` methods to `element` class
+
 # 08-03-2025 :
 - added `get_element_by_query` function
 - added `get_element_by_query` method to `element` class